+++ rust/theoros/src/servers/api/mod.rs
@@ -1,6 +1,7 @@
+pub mod docs;
 pub mod router;
 
-use std::{net::SocketAddr, path::PathBuf};
+use std::net::SocketAddr;
 
 use anyhow::{Context, Result};
 use router::api_router;
@@ -9,10 +10,8 @@ use tower_http::{
     cors::CorsLayer,
     trace::{DefaultMakeSpan, TraceLayer},
 };
-use utoipa::OpenApi;
-use utoipauto::utoipauto;
 
-use crate::{config::Config, AppState};
+use crate::{config::Config, servers::api::docs::ApiDoc, AppState};
 
 #[tracing::instrument(skip(config, state))]
 pub fn run_api_server(config: &Config, state: AppState) -> Result<JoinHandle<Result<()>>> {
@@ -38,27 +37,4 @@ pub fn run_api_server(config: &Config, state: AppState) -> Result<JoinHandle<Res
     });
 
     Ok(handle)
-}
-
-#[utoipauto(paths = "./theoros/src")]
-#[derive(OpenApi)]
-#[openapi(
-        tags(
-            (name = "theoros", description = "Theoros Pragma Consultant")
-        )
-    )]
-pub struct ApiDoc;
-
-impl ApiDoc {
-    pub fn generate_openapi_json(output_path: PathBuf) -> Result<()> {
-        let json = ApiDoc::openapi().to_json()?;
-        if let Some(parent) = output_path.parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-        let file_path = output_path.join("openapi.json");
-        tracing::info!("Saving OpenAPI specs to {} ....", file_path.as_path().display());
-        std::fs::write(file_path, json)?;
-        tracing::info!("OpenAPI specs saved!");
-        Ok(())
-    }
 }
